Moving fleet to spot for simplicity
